fix: move module to correct folder
