+++ app/(logged-in)/org/[slug]/dashboard/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Button } from '@/components/ui/button';
-import { CurrencyConverter } from '@/components/engine/currency-converter';
+import { CurrencyConverter } from './components/currency-converter';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
